refactor(wrapped): extract all story-card components into _components
Move the ~750-line block of presentational story cards (PosterMosaic,
StoryChrome, Intro/Total/Format/Genres/Creator/Standout/Picks/Outro
cards, FormatRow, TStory, ChromeProps, GENRE_TILE_GRADIENTS) verbatim
into _components/story-cards.tsx. The two remaining page<->card shared
symbols (STEP_DURATION_MS, stepBackgrounds) moved to _lib/story.ts so
both sides import them. Pruned 9 now-unused lucide icons from the page.
tsc caught every shared-symbol seam; behavior-identical. wrapped/page
1443->683 LOC.
